[CIVIC-207] Slider / Slide - Code Remediation.
