+++ docroot/themes/custom/civic/civic-library/components/00-base/base.stories.js
@@ -10,9 +10,6 @@ import Checkbox from '../01-atoms/checkbox/checkbox.twig';
 import Radio from '../01-atoms/radio/radio.twig';
 import FormElement from '../03-organisms/form-element/form-element.twig';
 import Label from '../01-atoms/label/label.twig';
-import Slide from '../03-organisms/slider/slide.twig';
-import Tag from '../01-atoms/tag/tag.twig';
-import Button from '../01-atoms/button/button.twig';
 
 export const getSlots = (names) => {
   const showSlots = boolean('Show story-slots', false, 'Slots');
@@ -150,59 +147,4 @@ export const randomFormElements = (count, theme, rand) => {
   }
 
   return formElements;
-};
-
-export const randomTagsComponent = (count, theme) => {
-  const tags = [];
-  for (let i = 0; i < count; i++) {
-    tags.push(Tag({
-      theme,
-      text: randomString(getRandomInt(3, 8)),
-    }));
-  }
-  return tags;
-};
-
-export const randomButtonsComponent = (count, theme) => {
-  const tags = [];
-  for (let i = 0; i < count; i++) {
-    tags.push(Button({
-      theme,
-      kind: 'link',
-      text: randomString(getRandomInt(3, 8)),
-      type: ['primary', 'secondary', 'tertiary'][getRandomInt(0, 2)],
-      url: 'https://www.salsadigital.com.au',
-    }));
-  }
-  return tags;
-};
-
-export const randomSlidesComponent = (count, theme, rand, template) => {
-  const slides = [];
-
-  const inverseTheme = theme === 'dark' ? 'light' : 'dark';
-
-  for (let i = 0; i < count; i++) {
-    const contentTop = (template && template.content_top) ? template.content_top : randomTagsComponent(getRandomInt(0, 4)).join(' ');
-    const title = (template && template.title) ? template.title : `Title ${i + 1}${rand ? ` ${randomString(getRandomInt(5, 30))}` : ''}`;
-    const summary = (template && template.summary) ? template.summary : `Summary ${i + 1}${rand ? ` ${randomString(getRandomInt(5, 250))}` : ''}`;
-    const links = (template && template.links) ? template.links : randomButtonsComponent(getRandomInt(0, 4), inverseTheme).join('');
-    const image = (template && template.image) ? template.image : {
-      src: demoImage(),
-      alt: randomText(4),
-    };
-    const contentBottom = (template && template.content_bottom) ? template.content_bottom : '';
-    const attributes = (template && template.attributes) ? template.attributes : 'data-component-civic-slider-slide';
-    slides.push(Slide({
-      theme,
-      content_top: contentTop,
-      title,
-      summary,
-      links,
-      image,
-      content_bottom: contentBottom,
-      attributes,
-    }));
-  }
-  return slides;
 };
